requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
paramiko==3.1.0
pycparser==2.21
PyNaCl==1.5.0
pyserial==3.5
python-engineio==4.7.1
python-socketio==5.9.0
simple-websocket==0.10.1
six==1.16.0
SQLAlchemy==2.0.15
telnetlib3==2.0.2
typing_extensions==4.6.2
Werkzeug==2.3.4
wsproto==1.2.0
